Add files via upload
Our project is centered on consumer analytics and the presentation of food alternatives on applications. Understanding how customers make their decisions is critical in today’s environment, especially with the development of meal delivery apps and the convenience they provide. According to Business Insider, the meal delivery sector will be worth $161.7 billion by 2023. As a result, we wanted to investigate how specific app features, such as a suggestion badge, influence the consumer’s decision-making process. Additionally, we intend to explore how this differs by gender, budget choices, and age.